models/requirements.txt
tensorflow==2.18.0
keras==3.7.0
keras-tuner==1.4.7
numpy==1.26.4
pandas==2.2.1
scikit-learn==1.4.1.post1
matplotlib==3.8.3
seaborn==0.13.2
plotly==5.20.0
opencv-python==4.9.0.80
Pillow==10.2.0
Flask==3.0.2
Flask-CORS==4.0.0
mlflow==2.11.1
tqdm==4.66.2
PyYAML==6.0.1
kaggle==1.6.3